[k1b] Enhancement: Refactoring Core Interface
