+++ src/hal/arch/cluster/k1b-cluster/cluster.c
@@ -22,17 +22,43 @@
  * SOFTWARE.
  */
 
-
+/* Must come first. */
 #define __NEED_HAL_CLUSTER
+
 #include <nanvix/hal/cluster.h>
+#include <HAL/hal/core/legacy.h>
+#include <HAL/hal/board/boot_args.h>
+#include <mOS_common_types_c.h>
+#include <mOS_vcore_u.h>
 #include <nanvix/const.h>
 #include <nanvix/klib.h>
-#include <HAL/hal/core/legacy.h>
-#include <mOS_common_types_c.h>
 
 /* Import definitions. */
 EXTERN NORETURN void kmain(int, const char *[]);
-EXTERN void _do_syscall(int, int, int, int, int, int, int, int);
+
+/**
+ * @brief Cores table.
+ */
+PUBLIC struct coreinfo ALIGN(K1B_CACHE_LINE_SIZE) cores[K1B_CLUSTER_NUM_CORES] = {
+	{ TRUE,  CORE_RUNNING,   0, NULL, K1B_SPINLOCK_UNLOCKED }, /* Master Core   */
+	{ FALSE, CORE_RESETTING, 0, NULL, K1B_SPINLOCK_LOCKED   }, /* Slave Core 1  */
+	{ FALSE, CORE_RESETTING, 0, NULL, K1B_SPINLOCK_LOCKED   }, /* Slave Core 2  */
+	{ FALSE, CORE_RESETTING, 0, NULL, K1B_SPINLOCK_LOCKED   }, /* Slave Core 3  */
+#if defined(__node__)
+	{ FALSE, CORE_RESETTING, 0, NULL, K1B_SPINLOCK_LOCKED   }, /* Slave Core 4  */
+	{ FALSE, CORE_RESETTING, 0, NULL, K1B_SPINLOCK_LOCKED   }, /* Slave Core 5  */
+	{ FALSE, CORE_RESETTING, 0, NULL, K1B_SPINLOCK_LOCKED   }, /* Slave Core 6  */
+	{ FALSE, CORE_RESETTING, 0, NULL, K1B_SPINLOCK_LOCKED   }, /* Slave Core 7  */
+	{ FALSE, CORE_RESETTING, 0, NULL, K1B_SPINLOCK_LOCKED   }, /* Slave Core 8  */
+	{ FALSE, CORE_RESETTING, 0, NULL, K1B_SPINLOCK_LOCKED   }, /* Slave Core 9  */
+	{ FALSE, CORE_RESETTING, 0, NULL, K1B_SPINLOCK_LOCKED   }, /* Slave Core 10 */
+	{ FALSE, CORE_RESETTING, 0, NULL, K1B_SPINLOCK_LOCKED   }, /* Slave Core 11 */
+	{ FALSE, CORE_RESETTING, 0, NULL, K1B_SPINLOCK_LOCKED   }, /* Slave Core 12 */
+	{ FALSE, CORE_RESETTING, 0, NULL, K1B_SPINLOCK_LOCKED   }, /* Slave Core 13 */
+	{ FALSE, CORE_RESETTING, 0, NULL, K1B_SPINLOCK_LOCKED   }, /* Slave Core 14 */
+	{ FALSE, CORE_RESETTING, 0, NULL, K1B_SPINLOCK_LOCKED   }, /* Slave Core 15 */
+#endif
+};
 
 /*============================================================================*
  * k1b_stack_setup()                                                          *
@@ -61,51 +87,54 @@ PRIVATE inline void k1b_stack_setup(void)
 }
 
 /*============================================================================*
- * k1b_core_setup()                                                           *
+ * k1b_cluster_master_setup()                                                 *
  *============================================================================*/
 
 /**
- * The k1b_core_setup() function initializes all architectural
- * structures of the underlying core. It setups the Interrupt Vector
- * Table (IVT) and the Memory Management Unit (MMU) tables.
+ * @brief Initializes the master core.
  *
- * @todo We should move this to the core module.
- * @bug We cannot move this out of this file due to somelinking problem.
+ * The k1b_cluster_master_setup() function initializes the underlying
+ * master core.  It setups the stack, retrieves the boot arguments,
+ * initializes architectural structures and then call the kernel main
+ * function.
+ *
+ * @note This function does not return.
  *
  * @author Pedro Henrique Penna
  */
-PUBLIC void k1b_core_setup(void)
+PRIVATE NORETURN void k1b_cluster_master_setup(void)
 {
-	kprintf("[hal] booting up core");
+	k1_boot_args_t args;
 
-	k1b_ivt_setup(
-		(k1b_hwint_handler_fn) k1b_do_hwint,
-		(k1b_swint_handler_fn) _do_syscall,
-		(k1b_excp_handler_fn) _do_excp
-	);
+	k1b_stack_setup();
 
-	k1b_cluster_mem_setup();
+	get_k1_boot_args(&args);
+
+	k1b_cluster_setup();
+
+	kmain(args.argc, (const char **)args.argv);
 }
 
 /*============================================================================*
- * k1b_slave_setup()                                                          *
+ * k1b_cluster_slave_setup()                                                          *
  *============================================================================*/
 
 /**
  * @brief Initializes a slave core.
  *
- * The k1b_slave_setup() function initializes the underlying slave
- * core.  It setups the stack and then call the kernel main function.
- * Architectural structures are initialized by the master core and
- * registered later on, when the slave core is started effectively.
+ * The k1b_cluster_slave_setup() function initializes the underlying
+ * slave core.  It setups the stack and then call the kernel main
+ * function.  Architectural structures are initialized by the master
+ * core and registered later on, when the slave core is started
+ * effectively.
  *
  * @note This function does not return.
  *
- * @see k1b_core_setup() and k1b_master_setup().
+ * @see k1b_cluster_setup() and k1b_cluster_master_setup().
  *
  * @author Pedro Henrique Penna
  */
-PUBLIC NORETURN void k1b_slave_setup(void)
+PUBLIC NORETURN void k1b_cluster_slave_setup(void)
 {
 	k1b_stack_setup();
 
@@ -117,32 +146,20 @@ PUBLIC NORETURN void k1b_slave_setup(void)
 }
 
 /*============================================================================*
- * k1b_master_setup()                                                         *
+ * k1b_cluster_setup()                                                        *
  *============================================================================*/
 
 /**
- * @brief Initializes the master core.
- *
- * The k1b_master_setup() function initializes the underlying
- * master core.  It setups the stack, retrieves the boot arguments,
- * initializes architectural structures and then call the kernel main
- * function.
- *
- * @note This function does not return.
+ * @todo TODO provide a detailed description for this function.
  *
  * @author Pedro Henrique Penna
  */
-PRIVATE NORETURN void k1b_master_setup(void)
+PUBLIC void k1b_cluster_setup(void)
 {
-	k1_boot_args_t args;
-
-	k1b_stack_setup();
-
-	get_k1_boot_args(&args);
+	kprintf("[hal] booting up cluster...");
 
 	k1b_core_setup();
-
-	kmain(args.argc, (const char **)args.argv);
+	k1b_cluster_mem_setup();
 }
 
 /*============================================================================*
@@ -165,7 +182,7 @@ PUBLIC void SECTION_TEXT NORETURN _do_slave_pe(uint32_t oldsp)
 {
 	UNUSED(oldsp);
 
-	k1b_slave_setup();
+	k1b_cluster_slave_setup();
 }
 
 /*============================================================================*
@@ -188,5 +205,5 @@ PUBLIC void SECTION_TEXT NORETURN _do_master_pe(uint32_t oldsp)
 {
 	UNUSED(oldsp);
 
-	k1b_master_setup();
+	k1b_cluster_master_setup();
 }
